+++ packages/bsb-core/bsb/cli/__init__.py
@@ -3,9 +3,10 @@ import inspect
 import json
 import sys
 
+from bsb_otel.tracer import get_bsb_tracer
+
 from .._contexts import get_cli_context, reset_cli_context
 from ..exceptions import CommandError, DryrunError
-from ..profiling import _telemetry_trace
 from .commands import load_root_command
 
 
@@ -30,7 +31,7 @@ def handle_command(command, dryrun=False, exit=False):
         for action in namespace.internal_action_list or ():
             action(namespace)
     if not dryrun or _can_dryrun(namespace.handler, namespace):
-        with _telemetry_trace(
+        with get_bsb_tracer("bsb-core").trace(
             "cli",
             attributes={
                 "bsb.cli_command": command,
@@ -41,7 +42,6 @@ def handle_command(command, dryrun=False, exit=False):
                     }
                 ),
             },
-            broadcast=True,
         ):
             namespace.handler(namespace, dryrun=dryrun)
     else:  # pragma: nocover