+++ packages/bsb-nest/docs/conf.py
@@ -41,6 +41,10 @@ autodoc_typehints = "both"
 templates_path = ["_templates"]
 exclude_patterns = ["_build", "Thumbs.db", ".DS_Store"]
 
+nitpick_ignore = [
+    ("py:class", "bsb_nest._kernel_proxy.NestModelTypeHandler"),
+]
+
 
 # -- Options for HTML output -------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#options-for-html-output