+++ libs/arborize/CHANGELOG.md
@@ -1,3 +1,13 @@
+## 7.4.0 (2026-05-29)
+
+### 🚀 Features
+
+- Mature OpenTelemetry integration ([#223](https://github.com/dbbs-lab/bsb/pull/223), [#1](https://github.com/dbbs-lab/bsb/issues/1))
+
+### ❤️ Thank You
+
+- Robin De Schepper
+
 ## 7.3.1 (2026-04-16)
 
 ### 🩹 Fixes